+++ src/BlazorUI/Bit.BlazorUI.Extras/Components/DataGrid/BitDataGrid.razor
@@ -5,7 +5,7 @@
     @{
         StartCollectingColumns();
     }
-    @ChildContent
+    @(Columns ?? ChildContent)
     <BitDataGridDefer>
         @{
             FinishCollectingColumns();
@@ -26,11 +26,22 @@
                                 TItem="(int RowIndex, TGridItem Data)"
                                 ItemsProvider="@ProvideVirtualizedItems"
                                 ItemContent="@(item => builder => RenderRow(builder, item.RowIndex, item.Data))"
-                                Placeholder="@(placeholderContext => builder => RenderPlaceholderRow(builder, placeholderContext))" />
+                    Placeholder="@(placeholderContext => builder => RenderPlaceholderRow(builder, placeholderContext))" />
                 }
                 else
                 {
-                    @_renderNonVirtualizedRows
+                    if (IsLoading && LoadingTemplate is not null)
+                    {
+                        <tr>
+                            <td colspan="100">
+                                @LoadingTemplate
+                            </td>
+                        </tr>
+                    }
+                    else
+                    {
+                        @_renderNonVirtualizedRows
+                    }
                 }
             </tbody>
         </table>
